Move the configuration machinery in a stand alone file
